services/face-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
opencv-python-headless
dlib
pytest
pytest-asyncio
flake8
qdrant-client
aio-pika
insightface
onnxruntime
python-dotenv